requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
pytz>=2024.1
websockets==12.0
python-dotenv==1.0.0
httpx>=0.25.2
loguru==0.7.2
